refactor: extract recipe schema contracts
